feature: Spring Security 커뮤니티 부분 권한 체크를 한다.
