+++ src/test/java/cbnu/io/cbnuswalrami/common/configuration/container/AcceptanceTestBase.java
@@ -1,0 +1,48 @@
+package cbnu.io.cbnuswalrami.common.configuration.container;
+
+import io.restassured.RestAssured;
+import io.restassured.builder.RequestSpecBuilder;
+import io.restassured.specification.RequestSpecification;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.restdocs.RestDocumentationContextProvider;
+import org.springframework.restdocs.RestDocumentationExtension;
+
+import static io.restassured.RestAssured.baseURI;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
+import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
+
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@ExtendWith({RestDocumentationExtension.class})
+public abstract class AcceptanceTestBase extends DatabaseTestBase{
+
+    public static final String BASE_URL = "http://localhost";
+
+    @LocalServerPort
+    private int port = 8080;
+
+    protected RequestSpecification documentationSpec = new RequestSpecBuilder()
+            .setPort(port)
+            .setBaseUri(BASE_URL)
+            .build();
+
+    @BeforeAll
+    public void configureRestAssured() {
+        RestAssured.port = port;
+        baseURI = BASE_URL;
+    }
+
+
+    @BeforeEach
+    public void setUpRestDocs(RestDocumentationContextProvider restDocumentation) {
+        this.documentationSpec = new RequestSpecBuilder()
+                .addFilter(documentationConfiguration(restDocumentation)
+                        .operationPreprocessors()
+                        .withRequestDefaults(prettyPrint())
+                        .withResponseDefaults(prettyPrint()))
+                .build();
+    }
+}
