+++ src/test/java/cbnu/io/cbnuswalrami/test/acceptance/approval/ChangeApprovalAcceptanceTest.java
@@ -1,61 +1,26 @@
 package cbnu.io.cbnuswalrami.test.acceptance.approval;
 
 import cbnu.io.cbnuswalrami.business.core.domon.member.entity.Member;
-import cbnu.io.cbnuswalrami.common.configuration.container.DatabaseTestBase;
+import cbnu.io.cbnuswalrami.common.configuration.container.AcceptanceTestBase;
 import cbnu.io.cbnuswalrami.test.helper.fixture.member.MemberFixture;
 import cbnu.io.cbnuswalrami.test.helper.fixture.member.SignupFixture;
-import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
-import io.restassured.specification.RequestSpecification;
 import org.junit.jupiter.api.*;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.restdocs.payload.ResponseFieldsSnippet;
-import org.springframework.restdocs.restassured3.RestAssuredRestDocumentation;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import static cbnu.io.cbnuswalrami.test.helper.util.ApiDocumentUtils.*;
-import static io.restassured.RestAssured.baseURI;
 import static io.restassured.RestAssured.given;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 
 @DisplayName("유저 승인기능 인수 테스트")
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@ExtendWith({RestDocumentationExtension.class, SpringExtension.class})
-public class ChangeApprovalAcceptanceTest extends DatabaseTestBase {
-
-    public static final String BASE_URL = "http://localhost";
-    @LocalServerPort
-    private int port = 8080;
+public class ChangeApprovalAcceptanceTest extends AcceptanceTestBase {
 
     @Autowired
     private SignupFixture signupFixture;
-
-
-    private RequestSpecification documentationSpec = new RequestSpecBuilder().setPort(port).setBaseUri(BASE_URL).build();
-
-    @BeforeAll
-    public void configureRestAssured() {
-        RestAssured.port = port;
-        baseURI = BASE_URL;
-    }
-
-
-    @BeforeEach
-    public void setUpRestDocs(RestDocumentationContextProvider restDocumentation) {
-        this.documentationSpec = new RequestSpecBuilder()
-                .addFilter(RestAssuredRestDocumentation.documentationConfiguration(restDocumentation))
-                .build();
-    }
-
 
     @DisplayName("회원가입한 유저가 있을 떄 어드민 계정은 해당 유저를 승인한다.")
     @Test
@@ -76,8 +41,6 @@ public class ChangeApprovalAcceptanceTest extends DatabaseTestBase {
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .filter(document(
                                 "change-approval",
-                                getDocumentRequest(),
-                                getDocumentResponse(),
                                 getResponseFieldsSnippet()
                         )
                 )
