+++ src/test/java/cbnu/io/cbnuswalrami/test/acceptance/notification/NotificationCursorListTest.java
@@ -1,26 +1,15 @@
 package cbnu.io.cbnuswalrami.test.acceptance.notification;
 
-import cbnu.io.cbnuswalrami.common.configuration.container.DatabaseTestBase;
+import cbnu.io.cbnuswalrami.common.configuration.container.AcceptanceTestBase;
 import cbnu.io.cbnuswalrami.test.helper.fixture.notification.NotificationFixture;
-import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
-import io.restassured.specification.RequestSpecification;
 import org.junit.jupiter.api.*;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.restdocs.payload.ResponseFieldsSnippet;
 import org.springframework.restdocs.request.RequestParametersSnippet;
-import org.springframework.restdocs.restassured3.RestAssuredRestDocumentation;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import static cbnu.io.cbnuswalrami.test.helper.util.ApiDocumentUtils.getDocumentRequest;
-import static cbnu.io.cbnuswalrami.test.helper.util.ApiDocumentUtils.getDocumentResponse;
 import static io.restassured.RestAssured.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
@@ -29,35 +18,10 @@ import static org.springframework.restdocs.request.RequestDocumentation.requestP
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 
 @DisplayName("공지사항 리스트 조회 API 인수 테스트")
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@ExtendWith({RestDocumentationExtension.class, SpringExtension.class})
-public class NotificationCursorListTest extends DatabaseTestBase {
-
-        public static final String BASE_URL = "http://localhost";
-        @LocalServerPort
-        private int port = 8080;
+public class NotificationCursorListTest extends AcceptanceTestBase {
 
         @Autowired
         private NotificationFixture notificationFixture;
-
-        private RequestSpecification documentationSpec = new RequestSpecBuilder()
-        .setPort(port)
-        .setBaseUri(BASE_URL)
-        .build();
-
-        @BeforeAll
-        public void configureRestAssured() {
-        RestAssured.port = port;
-        baseURI = BASE_URL;
-        }
-
-
-        @BeforeEach
-        public void setUpRestDocs(RestDocumentationContextProvider restDocumentation) {
-        this.documentationSpec = new RequestSpecBuilder()
-        .addFilter(RestAssuredRestDocumentation.documentationConfiguration(restDocumentation))
-        .build();
-        }
 
         @DisplayName("공지사항 리스트를 조회한다")
         @Test
@@ -72,8 +36,6 @@ public class NotificationCursorListTest extends DatabaseTestBase {
                         .filter(
                                 document(
                                         "get-notifications",
-                                        getDocumentRequest(),
-                                        getDocumentResponse(),
                                         getRequestParametersSnippet(),
                                         getResponseFieldsSnippet()
                                 )
