+++ src/test/java/cbnu/io/cbnuswalrami/test/acceptance/signup/SignupAcceptanceTest.java
@@ -1,27 +1,18 @@
 package cbnu.io.cbnuswalrami.test.acceptance.signup;
 
-import cbnu.io.cbnuswalrami.common.configuration.container.DatabaseTestBase;
+import cbnu.io.cbnuswalrami.common.configuration.container.AcceptanceTestBase;
 import cbnu.io.cbnuswalrami.test.helper.fake.s3.S3Configuration;
 import com.google.gson.JsonObject;
 import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
 import org.junit.jupiter.api.*;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.HttpStatus;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.restdocs.payload.RequestPartFieldsSnippet;
 import org.springframework.restdocs.payload.ResponseFieldsSnippet;
 import org.springframework.restdocs.request.RequestPartsSnippet;
-import org.springframework.restdocs.restassured3.RestAssuredRestDocumentation;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import static cbnu.io.cbnuswalrami.test.helper.util.ApiDocumentUtils.*;
-import static io.restassured.RestAssured.*;
 import static org.springframework.http.MediaType.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.restdocs.request.RequestDocumentation.partWithName;
@@ -29,34 +20,9 @@ import static org.springframework.restdocs.request.RequestDocumentation.requestP
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 
 @DisplayName("회원가입 기능 인수 테스트")
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@ExtendWith({RestDocumentationExtension.class, SpringExtension.class})
 @Import(S3Configuration.class)
-public class SignupAcceptanceTest extends DatabaseTestBase {
+public class SignupAcceptanceTest extends AcceptanceTestBase {
 
-    public static final String BASE_URL = "http://localhost";
-    @LocalServerPort
-    private int port = 8080;
-
-
-    private RequestSpecification documentationSpec = new RequestSpecBuilder()
-            .setPort(port)
-            .setBaseUri(BASE_URL)
-            .build();
-
-    @BeforeAll
-    public void configureRestAssured() {
-        RestAssured.port = port;
-        baseURI = BASE_URL;
-    }
-
-
-    @BeforeEach
-    public void setUpRestDocs(RestDocumentationContextProvider restDocumentation) {
-        this.documentationSpec = new RequestSpecBuilder()
-                .addFilter(RestAssuredRestDocumentation.documentationConfiguration(restDocumentation))
-                .build();
-    }
 
     @Test
     @DisplayName("회원가입 API 테스트")
@@ -72,8 +38,6 @@ public class SignupAcceptanceTest extends DatabaseTestBase {
                 .contentType(MULTIPART_FORM_DATA_VALUE)
                 .filter(document(
                         "users-post",
-                        getDocumentRequest(),
-                        getDocumentResponse(),
                         getRequestPartsSnippet(),
                         getRequestPartFieldsSnippet(),
                         getResponseFieldsSnippet()
