--- src/test/java/cbnu/io/cbnuswalrami/test/helper/util/ApiDocumentUtils.java
@@ -1,17 +1,0 @@
-package cbnu.io.cbnuswalrami.test.helper.util;
-
-import org.springframework.restdocs.operation.preprocess.OperationRequestPreprocessor;
-import org.springframework.restdocs.operation.preprocess.OperationResponsePreprocessor;
-
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
-
-public interface ApiDocumentUtils {
-
-    static OperationRequestPreprocessor getDocumentRequest() {
-        return preprocessRequest(prettyPrint());
-    }
-
-    static OperationResponsePreprocessor getDocumentResponse() {
-        return preprocessResponse(prettyPrint());
-    }
-}
